refactor: move state logic into custom hook useMacroCalculator
